Refactored to extract clear SPI. Addressed review comments
