+++ lib/addons/index.ts
@@ -10,7 +10,7 @@ export { MetricsServerAddOn } from './metrics-server'
 export { NginxAddOn } from './nginx'
 export { SSMAgentAddOn } from './ssm-agent'
 export { XrayAddOn } from './xray'
-
+``
 export class Constants {
     public static readonly SSP_ADDON = "ssp-addon";
 }
