+++ lib/spi/index.ts
@@ -1,0 +1,5 @@
+export * from './types';
+export * from './addon-contracts';
+export * from './cluster-contracts';
+export * from './team-contracts';
+
